Retrieve only projects user has access to during validation.
Validating causes the full public project list to be retrieved from the gitlab server.  When working on gitlab.com, this can cause eclipse to hang during validation as it retrieves every public project in the system. 

Changed the call from getProjects() to getMembershipProjects(),  which pulls only the projects that the user is a member of in some capacity - a small number of projects, rather than hundreds of thousands (millions?) of public repos available on gitlab.com.